+++ src/MissionManager/MissionCommandList.cc
@@ -12,7 +12,6 @@
 #include "Vehicle.h"
 #include "FirmwarePluginManager.h"
 #include "QGCApplication.h"
-#include "QGroundControlQmlGlobal.h"
 #include "JsonHelper.h"
 #include "MissionCommandUIInfo.h"