+++ src/MissionManager/MissionCommandUIInfo.cc
@@ -38,6 +38,7 @@ const char* MissionCommandUIInfo::_standaloneCoordinateJsonKey  = "standaloneCoo
 const char* MissionCommandUIInfo::_specifiesCoordinateJsonKey   = "specifiesCoordinate";
 const char* MissionCommandUIInfo::_specifiesAltitudeOnlyJsonKey = "specifiesAltitudeOnly";
 const char* MissionCommandUIInfo::_isLandCommandJsonKey         = "isLandCommand";
+const char* MissionCommandUIInfo::_isTakeoffCommandJsonKey      = "isTakeoffCommand";
 const char* MissionCommandUIInfo::_unitsJsonKey                 = "units";
 const char* MissionCommandUIInfo::_commentJsonKey               = "comment";
 const char* MissionCommandUIInfo::_advancedCategory             = "Advanced";
@@ -174,6 +175,15 @@ bool MissionCommandUIInfo::isLandCommand(void) const
     }
 }
 
+bool MissionCommandUIInfo::isTakeoffCommand(void) const
+{
+    if (_infoMap.contains(_isTakeoffCommandJsonKey)) {
+        return _infoMap[_isTakeoffCommandJsonKey].toBool();
+    } else {
+        return false;
+    }
+}
+
 void MissionCommandUIInfo::_overrideInfo(MissionCommandUIInfo* uiInfo)
 {
     // Override info values
@@ -209,7 +219,7 @@ bool MissionCommandUIInfo::loadJsonInfo(const QJsonObject& jsonObject, bool requ
     QStringList allKeys;
     allKeys << _idJsonKey << _rawNameJsonKey << _friendlyNameJsonKey << _descriptionJsonKey << _standaloneCoordinateJsonKey << _specifiesCoordinateJsonKey
             <<_friendlyEditJsonKey << _param1JsonKey << _param2JsonKey << _param3JsonKey << _param4JsonKey << _param5JsonKey << _param6JsonKey << _param7JsonKey
-            << _paramRemoveJsonKey << _categoryJsonKey << _specifiesAltitudeOnlyJsonKey << _isLandCommandJsonKey;
+            << _paramRemoveJsonKey << _categoryJsonKey << _specifiesAltitudeOnlyJsonKey << _isLandCommandJsonKey << _isTakeoffCommandJsonKey;
 
     // Look for unknown keys in top level object
     for (const QString& key: jsonObject.keys()) {
@@ -275,6 +285,9 @@ bool MissionCommandUIInfo::loadJsonInfo(const QJsonObject& jsonObject, bool requ
     if (jsonObject.contains(_isLandCommandJsonKey)) {
         _infoMap[_isLandCommandJsonKey] = jsonObject.value(_isLandCommandJsonKey).toBool();
     }
+    if (jsonObject.contains(_isTakeoffCommandJsonKey)) {
+        _infoMap[_isTakeoffCommandJsonKey] = jsonObject.value(_isTakeoffCommandJsonKey).toBool();
+    }
     if (jsonObject.contains(_friendlyEditJsonKey)) {
         _infoMap[_friendlyEditJsonKey] = jsonObject.value(_friendlyEditJsonKey).toVariant();
     }
@@ -304,6 +317,9 @@ bool MissionCommandUIInfo::loadJsonInfo(const QJsonObject& jsonObject, bool requ
         }
         if (!_infoAvailable(_isLandCommandJsonKey)) {
             _setInfoValue(_isLandCommandJsonKey, false);
+        }
+        if (!_infoAvailable(_isTakeoffCommandJsonKey)) {
+            _setInfoValue(_isTakeoffCommandJsonKey, false);
         }
         if (!_infoAvailable(_friendlyEditJsonKey)) {
             _setInfoValue(_friendlyEditJsonKey, false);