+++ src/MissionManager/MissionCommandUIInfo.h
@@ -96,6 +96,7 @@ private:
 /// specifiesCoordinate     bool    false       true: Command specifies a lat/lon/alt coordinate
 /// specifiesAltitudeOnly   bool    false       true: Command specifies an altitude only (no coordinate)
 /// standaloneCoordinate    bool    false       true: Vehicle does not fly through coordinate associated with command (exampl: ROI)
+/// isTakeoffCommand        bool    false       true: Command specifies a takeoff command (TAEKOFF, VTOL_TAKEOFF, ...)
 /// isLandCommand           bool    false       true: Command specifies a land command (LAND, VTOL_LAND, ...)
 /// friendlyEdit            bool    false       true: Command supports friendly editing dialog, false: Command supports 'Show all values" style editing only
 /// category                string  Advanced    Category which this command belongs to
@@ -120,6 +121,7 @@ public:
     Q_PROPERTY(bool     specifiesCoordinate     READ specifiesCoordinate    CONSTANT)
     Q_PROPERTY(bool     specifiesAltitudeOnly   READ specifiesAltitudeOnly  CONSTANT)
     Q_PROPERTY(bool     isLandCommand           READ isLandCommand          CONSTANT)
+    Q_PROPERTY(bool     isTakeoffCommand        READ isTakeoffCommand       CONSTANT)
     Q_PROPERTY(int      command                 READ intCommand             CONSTANT)
 
     MAV_CMD command(void) const { return _command; }
@@ -134,6 +136,7 @@ public:
     bool    specifiesCoordinate     (void) const;
     bool    specifiesAltitudeOnly   (void) const;
     bool    isLandCommand           (void) const;
+    bool    isTakeoffCommand        (void) const;
 
     /// Load the data in the object from the specified json
     ///     @param jsonObject Json object to load from
@@ -194,6 +197,7 @@ private:
     static const char* _specifiesCoordinateJsonKey;
     static const char* _specifiesAltitudeOnlyJsonKey;
     static const char* _isLandCommandJsonKey;
+    static const char* _isTakeoffCommandJsonKey;
     static const char* _unitsJsonKey;
     static const char* _commentJsonKey;    
     static const char* _advancedCategory;