+++ ChangeLog.md
@@ -7,6 +7,7 @@ Note: This file only contains high level features or important fixes.
 ### 4.0.6 - Not yet released
 
 * Analyze/Log Download - Fix download on mobile versions of QGC
+* Fly: Fix problems where Continue Mission and Change Altitude were not available after a Mission Pause.
 
 ### 4.0.5 - Stable